Incorporate workflow graph into website
