--- docs/index.html
@@ -1,218 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="utf-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <title>Workflow Graph (Graphviz preset + Cytoscape)</title>
-  <style>
-    :root {
-      color-scheme: light dark;
-      font-family: Inter, Segoe UI, Roboto, Helvetica, Arial, sans-serif;
-    }
-    body {
-      margin: 0;
-      display: grid;
-      grid-template-columns: 320px 1fr;
-      height: 100vh;
-    }
-    #sidebar {
-      border-right: 1px solid #8884;
-      padding: 12px;
-      overflow: auto;
-    }
-    #cy {
-      width: 100%;
-      height: 100%;
-      display: block;
-    }
-    input, select, button {
-      width: 100%;
-      margin: 0.3rem 0;
-      padding: 0.45rem;
-      box-sizing: border-box;
-    }
-    .muted {
-      opacity: 0.8;
-      font-size: 0.9rem;
-    }
-  </style>
-</head>
-<body>
-  <aside id="sidebar">
-    <h3>Workflow (Graphviz preset)</h3>
-    <div id="summary" class="muted">Loading...</div>
-    <label for="repoFilter">Repo filter</label>
-    <select id="repoFilter">
-      <option value="__all__">All repos</option>
-    </select>
-    <label for="searchBox">Search issues</label>
-    <input id="searchBox" type="text" placeholder="issue #, title..." />
-    <button id="resetView">Reset view</button>
-    <p class="muted">
-      Positions come from Graphviz (`dot -Tjson`) and are rendered with Cytoscape preset layout.
-      Click issue nodes to open GitHub.
-    </p>
-  </aside>
-  <main id="cy"></main>
-
-  <script src="https://unpkg.com/cytoscape@3.30.2/dist/cytoscape.min.js"></script>
-  <script>
-    async function init() {
-      const res = await fetch("workflow.json");
-      if (!res.ok) {
-        throw new Error(`Failed to load workflow.json: ${res.status}`);
-      }
-      const graph = await res.json();
-      const nodes = graph.elements.nodes || [];
-      const edges = graph.elements.edges || [];
-
-      const cy = cytoscape({
-        container: document.getElementById('cy'),
-        elements: [...nodes, ...edges],
-        style: [
-          {
-            selector: 'node',
-            style: {
-              'label': 'data(label)',
-              'font-size': 11,
-              'text-wrap': 'wrap',
-              'text-max-width': 220,
-              'text-valign': 'center',
-              'text-halign': 'center',
-              'shape': 'round-rectangle',
-            }
-          },
-          {
-            selector: 'node[type = "cluster"]',
-            style: {
-              'background-opacity': 0.06,
-              'border-width': 1.4,
-              'border-style': 'dashed',
-              'border-color': '#666',
-              'font-size': 12,
-              'font-weight': 600,
-              'text-valign': 'top',
-              'text-halign': 'center',
-              'text-wrap': 'wrap',
-              'text-max-width': 260,
-              'text-margin-y': -8,
-              'padding': '14px',
-            }
-          },
-          {
-            selector: 'node[type = "issue"]',
-            style: {
-              'background-opacity': 1,
-              'background-color': '#6DA6FF',
-              'font-size': 10,
-              'text-wrap': 'wrap',
-              // Match Graphviz layout: node bbox from `dot -Tjson` (gv_width/gv_height), not wrapped-label bounds.
-              'text-max-width': (ele) => {
-                const w = ele.data('gv_width');
-                if (w == null || Number.isNaN(Number(w))) return 220;
-                return Math.max(40, Number(w) - 24);
-              },
-              'width': (ele) => {
-                const w = ele.data('gv_width');
-                return (w != null && !Number.isNaN(Number(w))) ? Number(w) : 'label';
-              },
-              'height': (ele) => {
-                const h = ele.data('gv_height');
-                return (h != null && !Number.isNaN(Number(h))) ? Number(h) : 'label';
-              },
-              'padding': '12px',
-            }
-          },
-          {
-            selector: 'edge',
-            style: {
-              'curve-style': 'bezier',
-              'target-arrow-shape': 'triangle',
-              'width': 1.6,
-              'line-color': '#666',
-              'target-arrow-color': '#666',
-              'label': 'data(label)',
-              'font-size': 9,
-              'text-background-opacity': 1,
-              'text-background-color': '#fff',
-              'text-background-padding': 1,
-            }
-          },
-          {
-            selector: '.hidden',
-            style: {
-              'display': 'none'
-            }
-          }
-        ],
-        layout: {
-          name: 'preset',
-          fit: true,
-          padding: 30
-        }
-      });
-
-      const summary = document.getElementById('summary');
-      summary.textContent = `${graph.meta.issue_count} issues · ${graph.meta.edge_count} blocks edges`;
-
-      const repoFilter = document.getElementById('repoFilter');
-      for (const repo of graph.meta.repos || []) {
-        const option = document.createElement('option');
-        option.value = repo;
-        option.textContent = repo;
-        repoFilter.appendChild(option);
-      }
-
-      const applyFilters = () => {
-        const selectedRepo = repoFilter.value;
-        const query = document.getElementById('searchBox').value.trim().toLowerCase();
-        cy.elements().removeClass('hidden');
-
-        if (selectedRepo !== '__all__') {
-          cy.nodes('[type = "issue"]').forEach((node) => {
-            if (node.data('repo') !== selectedRepo) {
-              node.addClass('hidden');
-            }
-          });
-        }
-
-        if (query) {
-          cy.nodes('[type = "issue"]').forEach((node) => {
-            const hay = `${node.data('number')} ${node.data('title')}`.toLowerCase();
-            if (!hay.includes(query)) {
-              node.addClass('hidden');
-            }
-          });
-        }
-
-        cy.edges().forEach((edge) => {
-          if (edge.source().hasClass('hidden') || edge.target().hasClass('hidden')) {
-            edge.addClass('hidden');
-          }
-        });
-      };
-
-      repoFilter.addEventListener('change', applyFilters);
-      document.getElementById('searchBox').addEventListener('input', applyFilters);
-      document.getElementById('resetView').addEventListener('click', () => {
-        repoFilter.value = '__all__';
-        document.getElementById('searchBox').value = '';
-        cy.elements().removeClass('hidden');
-        cy.fit();
-      });
-
-      cy.on('tap', 'node[type = "issue"]', (evt) => {
-        const url = evt.target.data('url');
-        if (url) {
-          window.open(url, '_blank', 'noopener');
-        }
-      });
-    }
-
-    init().catch((error) => {
-      document.getElementById('summary').textContent = String(error);
-      console.error(error);
-    });
-  </script>
-</body>
-</html>
